# core/services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/gcp/roles": {
            "get": {
                "description": "List all IAM roles the project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List IAM Roles",
                "operationId": "gcp-list-role",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_gcp.RoleResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an IAM role for the project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update an IAM Role",
                "operationId": "gcp-update-role",
                "parameters": [
                    {
                        "description": "IAM role details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_gcp.RoleResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create an IAM role for the project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create an IAM Role",
                "operationId": "gcp-create-role",
                "parameters": [
                    {
                        "description": "IAM role details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.CreateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_gcp.RoleResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an IAM role for the project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete an IAM Role",
                "operationId": "gcp-delete-role",
                "parameters": [
                    {
                        "description": "IAM role name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.ActionRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "bool"
                        }
                    }
                }
            }
        },
...
        "/roles": {
            "get": {
                "description": "Get a list of IAM roles",
                "produces": [
                    "application/json"
                ],
                "summary": "List Roles",
                "operationId": "list-roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.RoleDetailResponse"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing IAM role's trust policy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update Role",
                "operationId": "update-role",
                "parameters": [
                    {
                        "description": "Role Name",
                        "name": "roleName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Updated Trust Policy",
                        "name": "trustPolicy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_requests_aws_iam.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.RoleDetailResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new IAM role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Role",
                "operationId": "create-role",
                "parameters": [
                    {
                        "description": "Role Name",
                        "name": "roleName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Trust Policy",
                        "name": "trustPolicy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_requests_aws_iam.CreateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.RoleDetailResponse"
                        }
                    }
                }
            }
        },
        "/roles/{roleName}": {
            "get": {
                "description": "Get IAM role details by role name",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Role by Name",
                "operationId": "get-role-by-name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role Name",
                        "name": "roleName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.RoleDetailResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an IAM role by role name",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Role",
                "operationId": "delete-role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role Name",
                        "name": "roleName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gitea_pcp-inariam_inariam_core_services_api_requests_aws_iam.CreateRoleRequest": {
            "type": "object",
            "required": [
                "name",
                "trust_policy"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "trust_policy": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "gitea_pcp-inariam_inariam_core_services_api_requests_aws_iam.UpdateRoleRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "new_name": {
                    "type": "string"
                },
                "trust_policy": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.ActionRoleRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.CreateRoleRequest": {
            "type": "object",
            "required": [
                "description",
                "name",
                "permissions",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stage": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "gitea_pcp-inariam_inariam_core_services_api_requests_gcp_iam.UpdateRoleRequest": {
            "type": "object",
            "required": [
                "description",
                "name",
                "permissions",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "gitea_pcp-inariam_inariam_core_services_api_responses_aws_iam.RoleDetailResponse": {
            "type": "object",
            "properties": {
                "role_arn": {
                    "type": "string"
                },
                "role_id": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
...
        "gitea_pcp-inariam_inariam_core_services_api_responses_gcp.RoleResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stage": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}